for #2567, refactor org.apache.shardingsphere.core.parse.sql.context
