+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/java/org/apache/shardingsphere/core/parse/sql/context/InsertValue.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.parse.sql.context.insertvalue;
+package org.apache.shardingsphere.core.parse.sql.context;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
